+++ apps/ui/src/Components/CommunityList/types.ts
@@ -1,0 +1,3 @@
+import { Community } from '@communecar/types';
+
+export type CommunityWithRides = Community & { rides: any[] };
